requirements.txt
aprs3==8.0.0.post1
aprslib==0.7.2
aprspy==0.4.0
attrs==25.3.0
ax253==0.1.5.post1
bitarray==3.5.0
bitstring==4.3.1
blinker==1.9.0
click==8.2.1
Flask==3.1.1
geographiclib==2.0
geopy==2.4.1
gevent==25.5.1
greenlet==3.2.3
gunicorn==23.0.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiss3==8.0.0
MarkupSafe==3.0.2
numpy==2.3.1
packaging==25.0
pyModeS==2.21.1
pyserial==3.5
pyserial-asyncio==0.6
pyzmq==27.0.0
setuptools==80.9.0
Werkzeug==3.1.3
zipp==3.23.0
zope.event==5.1
zope.interface==7.2
